1st_Place/DOC/src/requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==17.4.0
Automat==0.6.0
backcall==0.2.0
bleach==3.2.1
blinker==1.4
boto3==1.15.15
botocore==1.18.15
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
classification-models-3D==1.0.2
click==6.7
cloud-tpu-client==0.10
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.21
dataclasses
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
editdistance==0.5.3
efficientnet==1.1.1
efficientnet-3D==1.0.1
entrypoints==0.3
fasttext==0.9.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.0
google-api-core==1.22.4
google-api-python-client==1.8.0
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
image-classifiers==1.0.0
imageio==2.9.0
importlib-metadata==3.4.0
incremental==16.10.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.10
jmespath==0.10.0
joblib==0.17.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kaggle==1.5.8
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
lmdb==0.98
Markdown==3.3.3
MarkupSafe
matplotlib==3.3.2
mistune==0.8.4
ml-metrics==0.1.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
netifaces==0.10.4
networkx==2.5
nltk==3.4.5
notebook==6.1.4
numexpr==2.7.1
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
omegaconf==2.0.1rc4
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.4
PAM
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.5.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
python-debian==0.1.32
python-slugify==4.0.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
# PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.9.27
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.91
service-identity==16.0.0
six==1.15.0
sklearn==0.0
scikit-learn
numpy
pandas
slugify==0.0.1
smmap==3.0.4
ssh-import-id
python-systemd
tables==3.6.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.12.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.9.3
tornado==6.0.4
tqdm==4.50.2
traitlets==4.3.3
transformers==2.3.0
Twisted==17.9.0
typeguard==2.10.0
typing-extensions==3.7.4.3
pyufw
uritemplate==3.0.1
urllib3==1.22
validators==0.18.2
vecstack==0.4.0
vit-keras==0.0.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
zope.interface
